move types to NativeAlertManager
